refactor: Extract download screen components
- Extract `DownloadPager` and `DownloadPageContent` from `DownloadScreen`
- Isolate `DownloadScreen` navigation and `DownloadViewModel` interactions
- Simplify `DownloadPager` page ID mapping
