+++ app/src/main/kotlin/org/strigate/ferrot/presentation/MainNavHost.kt
@@ -13,10 +13,10 @@ import androidx.navigation.navArgument
 import org.strigate.ferrot.presentation.screen.AboutScreen
 import org.strigate.ferrot.presentation.screen.ArchivedScreen
 import org.strigate.ferrot.presentation.screen.CookiesScreen
-import org.strigate.ferrot.presentation.screen.DownloadScreen
 import org.strigate.ferrot.presentation.screen.GetCookiesScreen
 import org.strigate.ferrot.presentation.screen.SettingsScreen
 import org.strigate.ferrot.presentation.screen.UpdatesScreen
+import org.strigate.ferrot.presentation.screen.download.DownloadScreen
 import org.strigate.ferrot.presentation.screen.downloads.DownloadsScreen
 
 @Composable
